Rename NoneRenderer to BrowserRenderer
