+++ packages/rehype-mathjax/renderer/chtml.js
@@ -2,7 +2,7 @@ const {mathjax} = require('mathjax-full/js/mathjax')
 const {TeX} = require('mathjax-full/js/input/tex')
 const {CHTML} = require('mathjax-full/js/output/chtml')
 const {AllPackages} = require('mathjax-full/js/input/tex/AllPackages')
-const AbstractRenderer = require('./node')
+const AbstractRenderer = require('./base/node')
 
 class ChtmlRenderer extends AbstractRenderer {
   constructor(options) {
