+++ packages/rehype-mathjax/renderer/svg.js
@@ -2,7 +2,7 @@ const {mathjax} = require('mathjax-full/js/mathjax')
 const {TeX} = require('mathjax-full/js/input/tex')
 const {SVG} = require('mathjax-full/js/output/svg')
 const {AllPackages} = require('mathjax-full/js/input/tex/AllPackages')
-const AbstractRenderer = require('./node')
+const AbstractRenderer = require('./base/node')
 
 class SvgRenderer extends AbstractRenderer {
   constructor(options) {
